Added Dialog Screen with subTasks

+++ app/src/main/java/com/sinxn/myhabits/domain/use_case/task/UpdateTaskCompletedUseCase.kt
@@ -6,7 +6,7 @@ import javax.inject.Inject
 class UpdateTaskCompletedUseCase @Inject constructor(
     private val tasksRepository: TaskRepository,
 ) {
-    suspend operator fun invoke(taskId: Int, completed: Boolean) {
+    suspend operator fun invoke(taskId: Long, completed: Boolean) {
        // tasksRepository.completeTask(taskId, completed)
     }
 }
