+++ app/src/main/java/com/sinxn/myhabits/util/alarms/NotificationUtil.kt
@@ -21,7 +21,7 @@ fun NotificationManager.sendNotification(task: Task, context: Context, id: Int) 
     val completePendingIntent: PendingIntent =
         PendingIntent.getBroadcast(
             context,
-            task.id,
+            task.id.toInt(),
             completeIntent,
             PendingIntent.FLAG_IMMUTABLE or PendingIntent.FLAG_UPDATE_CURRENT
         )
